requirements.txt
Flask
Flask-Login>=0.3.0
Flask-Admin
Flask-SQLAlchemy
flask-mongoengine
wtforms
flask-wtf
markdown